refactor: move OAuthType to User.OAuthType
